pyproject.toml
[project]
name = "kob-erp-backend"
version = "0.1.0"
description = "KOB-ERP — standalone ERP backend (FastAPI + SQLAlchemy + PostgreSQL)"
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2.3",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "argon2-cffi>=23.1.0",
    "python-multipart>=0.0.20",
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "structlog>=24.4.0",
    "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-python-jose>=3.3.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["backend", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",  # pycodestyle, pyflakes
    "I",            # isort
    "UP",           # pyupgrade
    "B",            # bugbear
    "SIM",          # simplify
    "RUF",          # ruff-specific
    "TCH",          # type-checking
    "N",            # pep8-naming
]
ignore = ["E501"]  # line too long (ruff format handles it)

[tool.ruff.lint.per-file-ignores]
"backend/migrations/**" = ["N999"]
"tests/**" = ["S101"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = ["alembic.*", "celery.*"]
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning:passlib.*",
]

[tool.coverage.run]
source = ["backend"]
omit = ["backend/migrations/*", "backend/main.py"]
